external/webkit/LayoutTests/platform/android/plugins/npp-set-window-called-during-destruction-expected.txt

CONSOLE MESSAGE: line 13: PLUGIN: Success: NPP_SetWindow was called during plugin destruction
This tests whether a NPP_SetWindow is called during plugin destruction.
